+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-12 18:54:19
-// Commit: feat(test): add async test stream provider
+// Updated at: 2026-07-13 11:39:37
+// Commit: style(test): adjust design tokens, typography, and contrast in test
 
 class testComponentState {
-  final String commitRef = '2762';
-  final String lastMessage = 'feat(test): add async test stream provider';
-  final DateTime timestamp = DateTime.parse('2026-07-12T18:54:19+05:00');
+  final String commitRef = '2785';
+  final String lastMessage = 'style(test): adjust design tokens, typography, and contrast in test';
+  final DateTime timestamp = DateTime.parse('2026-07-13T11:39:37+05:00');
 }